+++ packages/cloud-cognitive/src/components/ExportModal/ExportModal.stories.js
@@ -1,5 +1,5 @@
 //
-// Copyright IBM Corp. 2020, 2020
+// Copyright IBM Corp. 2020, 2021
 //
 // This source code is licensed under the Apache-2.0 license found in the
 // LICENSE file in the root directory of this source tree.
@@ -12,6 +12,7 @@ import { pkg } from '../../settings';
 import { getStorybookPrefix } from '../../../config';
 import { ExportModal } from '.';
 import mdx from './ExportModal.mdx';
+import wait from '../../global/js/utils/wait';
 const storybookPrefix = getStorybookPrefix(pkg, ExportModal.displayName);
 
 export default {
@@ -27,13 +28,15 @@ export default {
 
 const defaultProps = {
   filename: 'Sample02.pdf',
+  iconDescription: 'close',
   inputLabel: 'File name',
-  modalHeading: 'Export',
   onRequestClose: () => {},
   onRequestSubmit: () => {},
   open: true,
+  preventCloseOnClickOutside: true,
   primaryButtonText: 'Export',
   secondaryButtonText: 'Cancel',
+  title: 'Export',
 };
 
 const Template = (args) => {
@@ -41,18 +44,16 @@ const Template = (args) => {
 };
 
 const TemplateWithState = (args) => {
-  const [open, setOpen] = useState(false);
+  const [open, setOpen] = useState(true);
   const [loading, setLoading] = useState(false);
   const [successful, setSuccessful] = useState(false);
   const [error, setError] = useState(false);
 
   const onSubmitHandler = async () => {
     setLoading(true);
-    await new Promise((res) => setTimeout(res, 1000));
-
-    if (args.exportSuccessful) setSuccessful(true);
+    await wait(1000);
+    if (args.successful) setSuccessful(true);
     else setError(true);
-
     setLoading(false);
   };
 
@@ -67,7 +68,7 @@ const TemplateWithState = (args) => {
       <ExportModal
         {...args}
         open={open}
-        onRequestClose={onCloseHandler}
+        onClose={onCloseHandler}
         onRequestSubmit={onSubmitHandler}
         loading={loading}
         successful={successful}
@@ -84,13 +85,13 @@ const TemplateWithState = (args) => {
 export const WithSuccessMessage = TemplateWithState.bind({});
 WithSuccessMessage.args = {
   ...defaultProps,
-  exportSuccessful: true,
+  successful: true,
 };
 
 export const WithErrorMessage = TemplateWithState.bind({});
 WithErrorMessage.args = {
   ...defaultProps,
-  exportSuccessful: false,
+  successful: false,
 };
 
 export const Standard = Template.bind({});
@@ -104,7 +105,7 @@ WithExtensionValidation.args = {
   validExtensions: ['pdf'],
   filename: '',
   invalidInputText: 'File must have valid extension .pdf',
-  modalBody: 'File must be exported in a PDF format.',
+  body: 'File must be exported in a PDF format.',
 };
 
 export const WithPreformattedExtensions = Template.bind({});